rename: schedule frontend 디렉토리 변경
