+++ frontend/src/app/login/ClientPage.tsx
@@ -2,7 +2,7 @@
 
 "use client";
 
-import { Button } from "@/components/ui/button";
+import { Button } from "@/components/schedule/ui/button";
 import { MessageCircle } from "lucide-react";
 import Link from "next/link";
