+++ frontend/src/types/schedule/schedule.ts
